requirements.txt
fastapi
uvicorn
python-dotenv
passlib
pymongo
bcrypt
python-jose
pyjwt
pydantic[email]
python-multipart
jinja2
python-dotenv